+++ Deep/Deep.csproj
@@ -50,8 +50,24 @@
       <Private>True</Private>
     </Reference>
     <Reference Include="Microsoft.CSharp" />
+    <Reference Include="Microsoft.Data.Edm, Version=5.6.0.0, Culture=neutral, PublicKeyToken=31bf3856ad364e35, processorArchitecture=MSIL">
+      <HintPath>..\packages\Microsoft.Data.Edm.5.6.0\lib\net40\Microsoft.Data.Edm.dll</HintPath>
+    </Reference>
+    <Reference Include="Microsoft.Data.OData, Version=5.6.0.0, Culture=neutral, PublicKeyToken=31bf3856ad364e35, processorArchitecture=MSIL">
+      <HintPath>..\packages\Microsoft.Data.OData.5.6.0\lib\net40\Microsoft.Data.OData.dll</HintPath>
+    </Reference>
+    <Reference Include="Microsoft.Data.Services, Version=5.6.0.0, Culture=neutral, PublicKeyToken=31bf3856ad364e35, processorArchitecture=MSIL">
+      <HintPath>..\packages\Microsoft.Data.Services.5.6.0\lib\net40\Microsoft.Data.Services.dll</HintPath>
+    </Reference>
+    <Reference Include="Microsoft.Data.Services.Client, Version=5.6.0.0, Culture=neutral, PublicKeyToken=31bf3856ad364e35, processorArchitecture=MSIL">
+      <HintPath>..\packages\Microsoft.Data.Services.Client.5.6.0\lib\net40\Microsoft.Data.Services.Client.dll</HintPath>
+    </Reference>
     <Reference Include="System.ComponentModel.DataAnnotations" />
+    <Reference Include="System.Data.Entity" />
     <Reference Include="System.Security" />
+    <Reference Include="System.Spatial, Version=5.6.0.0, Culture=neutral, PublicKeyToken=31bf3856ad364e35, processorArchitecture=MSIL">
+      <HintPath>..\packages\System.Spatial.5.6.0\lib\net40\System.Spatial.dll</HintPath>
+    </Reference>
     <Reference Include="System.Web.DynamicData" />
     <Reference Include="System.Web.Entity" />
     <Reference Include="System.Web.ApplicationServices" />
@@ -72,6 +88,7 @@
   </ItemGroup>
   <ItemGroup>
     <Content Include="LoginSystem\LoginService.svc" />
+    <Content Include="ProjectManagementSystem\ProjectService.svc" />
     <Content Include="Web.config" />
   </ItemGroup>
   <ItemGroup>
@@ -102,7 +119,9 @@
       <DependentUpon>LoginService.svc</DependentUpon>
     </Compile>
     <Compile Include="ProjectManagementSystem\IProjectService.cs" />
-    <Compile Include="ProjectManagementSystem\ProjectService.cs" />
+    <Compile Include="ProjectManagementSystem\ProjectService.svc.cs">
+      <DependentUpon>ProjectService.svc</DependentUpon>
+    </Compile>
     <Compile Include="Projects.cs">
       <DependentUpon>DeepModel.tt</DependentUpon>
     </Compile>
@@ -121,7 +140,9 @@
     </Compile>
   </ItemGroup>
   <ItemGroup>
+    <Folder Include="AnswerManagementSystem\" />
     <Folder Include="App_Data\" />
+    <Folder Include="QuestionManagementSystem\" />
   </ItemGroup>
   <ItemGroup>
     <Content Include="packages.config" />
