+++ Deep/DeepModel.edmx.diagram
@@ -5,13 +5,13 @@
     <!-- Diagram content (shape and connector positions) -->
     <edmx:Diagrams>
       <Diagram DiagramId="f8186d9ea69a43dd88a9e20cd3bd7af7" Name="Diagram1">
-        <EntityTypeShape EntityType="DB_A0BEB4_deepModel.AccessTokens" Width="1.5" PointX="0.75" PointY="0.75" IsExpanded="true" />
-        <EntityTypeShape EntityType="DB_A0BEB4_deepModel.Answers" Width="1.5" PointX="2.75" PointY="0.75" IsExpanded="true" />
-        <EntityTypeShape EntityType="DB_A0BEB4_deepModel.Questions" Width="1.5" PointX="2.75" PointY="3.75" IsExpanded="true" />
-        <EntityTypeShape EntityType="DB_A0BEB4_deepModel.Surveys" Width="1.5" PointX="4.75" PointY="0.75" IsExpanded="true" />
-        <EntityTypeShape EntityType="DB_A0BEB4_deepModel.Variants" Width="1.5" PointX="6.75" PointY="0.75" IsExpanded="true" />
-        <EntityTypeShape EntityType="DB_A0BEB4_deepModel.Users" Width="1.5" PointX="5.375" PointY="3.75" />
-        <EntityTypeShape EntityType="DB_A0BEB4_deepModel.Projects" Width="1.5" PointX="0.75" PointY="6.75" />
+        <EntityTypeShape EntityType="DB_A0BEB4_deepModel.AccessTokens" Width="1.5" PointX="0.75" PointY="0.75" />
+        <EntityTypeShape EntityType="DB_A0BEB4_deepModel.Answers" Width="1.5" PointX="2.75" PointY="0.75" />
+        <EntityTypeShape EntityType="DB_A0BEB4_deepModel.Projects" Width="1.5" PointX="0.75" PointY="3.75" />
+        <EntityTypeShape EntityType="DB_A0BEB4_deepModel.Questions" Width="1.5" PointX="2.75" PointY="3.75" />
+        <EntityTypeShape EntityType="DB_A0BEB4_deepModel.Surveys" Width="1.5" PointX="4.75" PointY="0.75" />
+        <EntityTypeShape EntityType="DB_A0BEB4_deepModel.Users" Width="1.5" PointX="4.75" PointY="3.75" />
+        <EntityTypeShape EntityType="DB_A0BEB4_deepModel.Variants" Width="1.5" PointX="6.75" PointY="0.75" />
       </Diagram>
     </edmx:Diagrams>
   </edmx:Designer>
