+++ Deep/packages.config
@@ -2,4 +2,9 @@
 <packages>
   <package id="EntityFramework" version="6.1.3" targetFramework="net451" />
   <package id="EntityFramework.ru" version="6.0.0" targetFramework="net451" />
+  <package id="Microsoft.Data.Edm" version="5.6.0" targetFramework="net451" />
+  <package id="Microsoft.Data.OData" version="5.6.0" targetFramework="net451" />
+  <package id="Microsoft.Data.Services" version="5.6.0" targetFramework="net451" />
+  <package id="Microsoft.Data.Services.Client" version="5.6.0" targetFramework="net451" />
+  <package id="System.Spatial" version="5.6.0" targetFramework="net451" />
 </packages>
